+++ src/frontend/web_application/src/services/protocols-config/index.js
@@ -14,7 +14,8 @@ export default {
     regexp: /^[a-z0-9.!#$%&*+=?_{}~-]+@([a-z0-9]+\.)?[a-z0-9][a-z0-9-]*\.[a-z]{2,60}$/,
   },
   twitter: {
-    regexp: /^@\w{1,15}$/,
+    // https://help.twitter.com/en/managing-your-account/twitter-username-rules
+    regexp: /^\w{1,15}$/,
   },
   facebook: {
   },