+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Disable draft form in case there is no selected identity
 - Facebook username is no more available in contact edition
+- In a draft, switch identity will change the protocol of all recipients as well
 
 ### Fixed